requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
black==22.8.0
    # via -r requirements.in
build==0.8.0
    # via pip-tools
click==8.1.3
    # via
    #   black
    #   pip-tools
django==3.2.9
    # via
    #   -r requirements.in
    #   django-js-asset
django-ckeditor==6.2.0
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-ckeditor
flake8==4.0.1
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via build
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytz==2022.2.1
    # via django
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   black
    #   build
    #   pep517
typing-extensions==4.3.0
    # via black
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
